增加64位兼容，并修改已知bug
1.支持64位编译
2.修复空间查询时使用非smID字段返回结果为空的bug
3.修复最佳路径分析返回结果为空的bug
4.修改rest服务回调的NotificationName
5.修复doc文档中的描述错误

+++ MapView/Map/RMTileCache.m
@@ -179,7 +179,7 @@
 			capacity =  value;
 			minimalPurge = MAX(1,capacity / 10);
 		} else 
-			RMLog(@"illegal value for capacity: %d", value);
+			RMLog(@"illegal value for capacity: %ld",(long)value);
 	}
 	
 	NSString* strategyStr = [cfg objectForKey:@"strategy"];
